[FEATURE] Allow manipulation of values while runtime for importers and migrators
Sometimes it is needed to manipulate values while runtime with an additional update query. Nevertheless at the end of the runtime those values could be overwritten again by a finalizing migration or import. Now you can add your new values to a query which will be respected at the final migration/import.