throw exception when exporting page 0
when exporting the page tree starting with page 0 the import fails with a misleading exception.
```
[ TypeError ]
Argument 1 passed to In2code\Migration\Utility\DatabaseUtility::isFieldExistingInTable() must be of the type string, int given, called in /var/www/html/private/typo3conf/ext/migration/Classes/Port/Import.php on line 438
```

The reason is mal-formed entry for the page:
```
"pages":[[false]]
```

To provide some convenience for the user an exception is thrown to help clearify that exporting page 0 or handling more than one tree at a time is not supported.