[TASK] Allow manipulation of values while migration with Queue class
In the past we introduces Queue class to set values while migrating. While runtime values could be manually read from Queue class. Now, we go a step further. If Queue was manipulated while runtime of a different element, this will be respected now at the end of a migration or import