[BUGFIX] Respect TCA rootLevel setting
Target pid is adjusted during imports, as the TCA rootLevel
setting is not considered the pid is e.g. adjusted for
sys_file_metadata, this will cause errors if not imported to
root page (=0).

By checking the TCA rootLevel setting of the imported tables
and skipping the adjustment for all tables that can only exist
in the root, this issue is fixed.

Resolves: #37